More refactor to create new furtherEvidence service to encapsulate logic so that we can reuse it easily
