+++ src/main/java/uk/gov/hmcts/reform/sscs/callback/handlers/IssueFurtherEvidenceHandler.java
@@ -10,29 +10,15 @@ import uk.gov.hmcts.reform.sscs.callback.CallbackHandler;
 import uk.gov.hmcts.reform.sscs.ccd.callback.Callback;
 import uk.gov.hmcts.reform.sscs.ccd.callback.CallbackType;
 import uk.gov.hmcts.reform.sscs.ccd.callback.DispatchPriority;
-import uk.gov.hmcts.reform.sscs.ccd.domain.EventType;
 import uk.gov.hmcts.reform.sscs.ccd.domain.SscsCaseData;
 import uk.gov.hmcts.reform.sscs.ccd.domain.SscsDocument;
-import uk.gov.hmcts.reform.sscs.ccd.service.CcdService;
-import uk.gov.hmcts.reform.sscs.docmosis.domain.Pdf;
-import uk.gov.hmcts.reform.sscs.idam.IdamService;
-import uk.gov.hmcts.reform.sscs.service.BulkPrintService;
-import uk.gov.hmcts.reform.sscs.service.CoverLetterService;
-import uk.gov.hmcts.reform.sscs.service.SscsDocumentService;
+import uk.gov.hmcts.reform.sscs.service.FurtherEvidenceService;
 
 @Service
 public class IssueFurtherEvidenceHandler implements CallbackHandler<SscsCaseData> {
 
     @Autowired
-    private CoverLetterService coverLetterService;
-    @Autowired
-    private SscsDocumentService sscsDocumentService;
-    @Autowired
-    private BulkPrintService bulkPrintService;
-    @Autowired
-    private CcdService ccdService;
-    @Autowired
-    private IdamService idamService;
+    private FurtherEvidenceService furtherEvidenceService;
 
     @Override
     public boolean canHandle(CallbackType callbackType, Callback<SscsCaseData> callback) {
@@ -55,29 +41,7 @@ public class IssueFurtherEvidenceHandler implements CallbackHandler<SscsCaseData
         if (!canHandle(callbackType, callback)) {
             throw new IllegalStateException("Cannot handle callback");
         }
-        SscsCaseData caseData = callback.getCaseDetails().getCaseData();
-        bulkPrintService.sendToBulkPrint(buildPdfsToBulkPrint(caseData), caseData);
-        setEvidenceIssuedFlagToYes(caseData);
-    }
-
-    private void setEvidenceIssuedFlagToYes(SscsCaseData caseData) {
-        sscsDocumentService.filterByDocTypeAndApplyAction(caseData.getSscsDocument(), APPELLANT_EVIDENCE,
-            doc -> doc.getValue().setEvidenceIssued("Yes"));
-        ccdService.updateCase(
-            caseData,
-            Long.valueOf(caseData.getCcdCaseId()),
-            EventType.UPDATE_CASE_ONLY.getCcdType(),
-            "Update case data only",
-            "Simply update case data with no callbacks afterwards",
-            idamService.getIdamTokens());
-    }
-
-    private List<Pdf> buildPdfsToBulkPrint(SscsCaseData caseData) {
-        List<Pdf> pdfsToBulkPrint = sscsDocumentService.getPdfsForGivenDocType(
-            caseData.getSscsDocument(), APPELLANT_EVIDENCE);
-        byte[] coverLetterContent = coverLetterService.generate609_97_OriginalSenderCoverLetter(caseData);
-        coverLetterService.appendCoverLetter(coverLetterContent, pdfsToBulkPrint);
-        return pdfsToBulkPrint;
+        furtherEvidenceService.issue(callback.getCaseDetails().getCaseData(), APPELLANT_EVIDENCE);
     }
 
     @Override
